1. If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include "function_pointers.h"
+#include <stdlib.h>
 
 /**
  * array_iterator - this function executes a function
@@ -17,7 +18,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int a;
+	size_t a;
 
 	if (array != NULL || action != NULL)
 	{
